Extract shared ReadingsRenderer module
Create ReadingsRenderer class to eliminate code duplication between
LiturgyOfTheDay and LiturgyOfAnyDay components:

- Add static frozen maps: readingLabels, massLabels, readingOrder
- Implement hasNestedSchemas, renderSingleReadings, renderReadings methods
- Add HTMLElement validation for container parameters
- Use Object.prototype.hasOwnProperty.call for safer property checks
- Both components now delegate readings rendering to the shared module
